Deleted .gitignore
It's not a good idea to add a .gitignore yet, especially since we haven't even figured out our system architecture, let alone what technology we want to use to implement said system architecture and how we want to implement it. This can easily be re-added once we have a more clear idea of the direction we want to take the project.